query_result: reimplement `strtobool`
Python 3.12 removes `distutils` module completely, as advised
in PEP 632. Migration notes [1] advises to reimplement `strtobool` as
there is no module that replaces it.

The implementation of the function is straighforward since the whole
logic is described in `distutils` API reference [2].

This commit reimplements this function to drop `distutils` dependency
completely.

[1]: https://peps.python.org/pep-0632/#migration-advice
[2]: https://docs.python.org/3.9/distutils/apiref.html#distutils.util.strtobool

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>