Cluster performance improvements
Speed up the computation for slots when initializing a cluster. After
profiling, this turned out to be very slow, when it does not have to be.
It does not make sense to recompute the same thing over and over
in a loop.

This change was originally part of redis-py's d1b4191f7a

Co-authored-by: Gabriel Erzse <gabriel.erzse@redis.com>
Co-authored-by: Chayim <chayim@users.noreply.github.com>
Signed-off-by: Salvatore Mesoraca <salvatore.mesoraca@aiven.io>