github: replace call for sleep with wait-for-it.sh
Previously integration tests invoked `sleep 10` to "give time to settle"
the environment. In a fast container, waiting for so long is not needed
and in a slow one it might not be enough.

This commit changes this line to use `./util/wait-for-it.sh` instead [1]
to check robustly that the port is being listened to.

Fixes #49.

[1] https://github.com/vishnubob/wait-for-it

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>