tests: make test_client_kill_filter_by_maxage more robust
Previously the test relied on the amount of clients that are currently
connected to the server. This is not robust as this can be affected by
the outside or a bad timing. This caused quite a lot of test case
failures.
The proper way to handle this is:
* Create a client and assign a distinguishable name to it
* Verify it's in `CLIENT LIST`
* Sleep for enough time for it to be killed by maxage we provide
* Verify it's not anymore in `CLIENT LIST`

This commit does exactly that.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>
Signed-off-by: Raphaël Vinot <raphael@vinot.info>