tests: fix floating-point error for georadius tests
The tests expected exactly "88.05060698409301" as Valkey georadius
result. However, on ARM64 machine the actual result is
"88.05060698338646". The difference is roughly 7e-10 and is not caused
by valkey-py, but rather a floating point error in georadius
calculations inside Valkey. This commit allows such an error to happen
without failing the test.

Signed-off-by: Mikhail Koviazin <mikhail.koviazin@aiven.io>